MXS-3404 Extend sqlite3 op-size from u8 to u16
To make it possible to have more tokens than 255.

Parsers operators (i.e. tokens) is one thing and opcodes
for the virtual machine of sqlite3 is another. Unfortunately
they are not completely separate, but some of the opcodes
in <build-directory>/opcodes.h are the same as the tokens in
<build-directory>/parse.h. And while the parser tokens are now
16-bit, the VM opcodes are 8-bit. However, this is probably not
a problem even if some of the parser tokens that are duplicated
in the opcodes are > 256 as we only use sqlite3 for parsing and
not for executing anything (on the sqlite3 VM).

+++ query_classifier/qc_sqlite/sqlite-src-3110100/src/expr.c
@@ -464,7 +464,7 @@ Expr *sqlite3ExprAlloc(
   pNew = sqlite3DbMallocRawNN(db, sizeof(Expr)+nExtra);
   if( pNew ){
     memset(pNew, 0, sizeof(Expr));
-    pNew->op = (u8)op;
+    pNew->op = (u16)op;
     pNew->iAgg = -1;
     if( pToken ){
       if( nExtra==0 ){
@@ -1544,7 +1544,7 @@ int sqlite3ExprIsInteger(Expr *p, int *pValue){
 ** TRUE.
 */
 int sqlite3ExprCanBeNull(const Expr *p){
-  u8 op;
+  u16 op;
   while( p->op==TK_UPLUS || p->op==TK_UMINUS ){ p = p->pLeft; }
   op = p->op;
   if( op==TK_REGISTER ) op = p->op2;
@@ -1574,7 +1574,7 @@ int sqlite3ExprCanBeNull(const Expr *p){
 ** answer.
 */
 int sqlite3ExprNeedsNoAffinityChange(const Expr *p, char aff){
-  u8 op;
+  u16 op;
   if( aff==SQLITE_AFF_BLOB ) return 1;
   while( p->op==TK_UPLUS || p->op==TK_UMINUS ){ p = p->pLeft; }
   op = p->op;
